Move allocators to separate files
